Move view-models to shared folder
