+++ app/backend/src/core/controllers/appraiser/handlers/TokenTransfers.ts
@@ -14,10 +14,8 @@ import {
   TransactionReceipt,
 } from "@/core/clients/blockchain/IBlockchainClient";
 import { mergeValues, ValueMapping } from "@/core/controllers/appraiser/types";
-import {
-  ValueRecord,
-  ValueType,
-} from "@/database/repositories/BlockValueRepository";
+import { ValueRecord } from "@/database/repositories/BlockValueRepository";
+import { ValueType } from "@/shared/api/viewModels/SyncStatusEndpoint";
 
 export class TokenTransferHandler extends BaseHandler {
   private readonly priceService: PriceService;
